requirements.txt
Flask>=2.3.0
requests>=2.31.0
googlemaps>=4.10.0
openrouteservice>=2.3.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.3.0
playsound>=1.3.0
geopy>=2.4.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
PyAudio>=0.2.11
Pillow>=10.0.0
